# docs/openapi.yml
info:
  title: IIS Server
  description: The implementation of the IIS server.
  version: 1.0.0
openapi: 3.0.0
components:
  schemas:
...
    Campaign:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        themeId:
          type: integer
...
    CampaignStepFull:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        active:
          type: boolean
...
    PageInfo:
      type: object
      properties:
        index:
          type: integer
        size:
          type: integer
        totalPages:
          type: integer
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
    PageIndex:
      name: pageIndex
      in: query
      required: true
      schema:
        type: integer
    PageSize:
      name: pageSize
      in: query
      required: true
      schema:
        type: integer
    Assigned:
      name: assigned
      in: query
      required: false
      schema:
        type: boolean
        description: Filter only those that are assigned to the current user.
...
paths:
...
  /campaigns:
    post:
      summary: Create a campaign.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                themeId:
                  type: integer
      responses:
        '200':
          description: Campaign created successfully.
    get:
      summary: Get campaigns by some criteria.
      parameters:
        - $ref: '#/components/parameters/PageIndex'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Assigned'
        - name: themeId
          in: query
          required: false
          schema:
            type: integer
      description: If 'assigned' is true, returns only campaigns assigned to the current user. If 'themeId' is provided, filters campaigns by the specified theme ID.
      responses:
        '200':
          description: List of campaigns.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Campaign'
                  page:
                    $ref: '#/components/schemas/PageInfo'
  /campaigns/{id}:
    delete:
      summary: Delete a campaign by ID.
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Campaign deleted successfully.
  /campaigns/{id}/inspect:
    get:
      summary: Inspect campaign by ID.
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Campaign deep information.
          content:
            application/json:
              schema:
                type: object
                properties:
                  campaign:
                    $ref: '#/components/schemas/Campaign'
                  assignedUserId:
                    type: integer
                    description: The ID of the user assigned to this campaign, if any.
  /campaigns/{id}/steps:
    get:
      summary: Get steps of a campaign by ID.
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: List of campaign steps.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/CampaignStepFull'
  /campaigns/{id}/assign:
    post:
      summary: Assign a user to a campaign.
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: integer
      responses:
        '200':
          description: Campaign assigned successfully.
  /campaigns/{id}/adduser:
    post:
      summary: Add a user to a campaign.
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: integer
      responses:
        '200':
          description: User added to campaign successfully.
  /campaigns/{id}/removeuser:
    post:
      summary: Remove a user from a campaign.
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: integer
      responses:
        '200':
          description: User removed from campaign successfully.
...